feat: Add support for early mixing of text in multimodal training
This commit modifies the `train.py` file to add support for early mixing of text in multimodal training. It introduces a new flag `early_mix_text` in the `DataArguments` class, which, when enabled, includes the conversation length in the `length_list` calculation even if there are no image or video modalities present. This change enhances the flexibility of the training process and allows for better handling of multimodal data.